feat: show different command palette actions depending on context
