+++ src/atoms/hooks/useActivePaneId.ts
@@ -1,0 +1,7 @@
+import { useAtomValue } from 'jotai';
+
+import { activePaneIdAtom } from '../core/activePane';
+
+export function useActivePaneId() {
+  return useAtomValue(activePaneIdAtom);
+}
